refactor: reorganizada pasta de utils
