+++ src/main/java/utils/terminal/tabelas/TablePrinter.java
@@ -1,4 +1,4 @@
-package utils;
+package utils.terminal.tabelas;
 
 import com.github.freva.asciitable.AsciiTable;
 import com.github.freva.asciitable.Column;
@@ -7,7 +7,6 @@ import com.github.freva.asciitable.ColumnData;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 public class TablePrinter {
